// FrontEnd/src/app/services/models/inverter.ts
/* tslint:disable */
/* eslint-disable */
export interface Inverter {
  frequency?: string;
  id?: number;
  image?: string;
  maximum_circuit_voltage?: number;
  minimal_circuit_current?: number;
  model?: string;
  nominal_power?: number;
  phase_number?: number;
  price?: number;
  quantity?: number;
  tension?: number;
  type?: string;
}